requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
ollama>=0.2.0
